Refactor reading/writing stable memory bytes
